+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/VideoPlayer.java
@@ -56,6 +56,15 @@ public abstract class VideoPlayer implements Messages.VideoPlayerInstanceApi {
     exoPlayer.prepare();
     exoPlayer.addListener(createExoPlayerEventListener(exoPlayer, surfaceProducer));
     setAudioAttributes(exoPlayer, options.mixWithOthers);
+
+    if (options.allowBackgroundPlayback) {
+        exoPlayer.setWakeMode(C.WAKE_MODE_NETWORK);
+        exoPlayer.setAudioAttributes(
+            new AudioAttributes.Builder()
+                .setUsage(C.USAGE_MEDIA)
+                .build(),
+            /* handleAudioFocus= */ !options.mixWithOthers);
+    }
   }
 
   public void setDisposeHandler(@Nullable DisposeHandler handler) {