+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/texture/TextureVideoPlayer.java
@@ -29,6 +29,7 @@ import io.flutter.view.TextureRegistry.SurfaceProducer;
 public final class TextureVideoPlayer extends VideoPlayer implements SurfaceProducer.Callback {
   // True when the ExoPlayer instance has a null surface.
   private boolean needsSurface = true;
+  @Nullable private final VideoPlayerOptions options;
   /**
    * Creates a texture video player.
    *
@@ -69,6 +70,7 @@ public final class TextureVideoPlayer extends VideoPlayer implements SurfaceProd
     super(events, mediaItem, options, surfaceProducer, exoPlayerProvider);
 
     surfaceProducer.setCallback(this);
+    this.options = options;
 
     Surface surface = surfaceProducer.getSurface();
     this.exoPlayer.setVideoSurface(surface);
@@ -100,8 +102,14 @@ public final class TextureVideoPlayer extends VideoPlayer implements SurfaceProd
 
   @RestrictTo(RestrictTo.Scope.LIBRARY)
   public void onSurfaceCleanup() {
-    exoPlayer.setVideoSurface(null);
-    needsSurface = true;
+     if (options == null || !options.allowBackgroundPlayback) {
+      savedStateDuring = ExoPlayerState.save(exoPlayer);
+      exoPlayer.release();
+    } else {
+      savedStateDuring = null;
+      exoPlayer.setVideoSurface(null);
+      needsSurface = true;
+    }
   }
 
   public void dispose() {