+++ packages/video_player/video_player_avfoundation/pigeons/messages.dart
@@ -52,6 +52,8 @@ abstract class AVFoundationVideoPlayerApi {
   TexturePlayerIds createForTextureView(CreationOptions creationOptions);
   @ObjCSelector('setMixWithOthers:')
   void setMixWithOthers(bool mixWithOthers);
+  @ObjCSelector('setAllowBackgroundPlayback:')
+  void setAllowBackgroundPlayback(bool allowBackgroundPlayback);
   @ObjCSelector('fileURLForAssetWithName:package:')
   String? getAssetUrl(String asset, String? package);
 }